+++ src/js/components/lightbox.js
@@ -66,7 +66,7 @@
 
         init: function() {
 
-            var $this = this, siblings = [];
+            var siblings = [];
 
             this.index    = 0;
             this.siblings = [];
@@ -352,7 +352,7 @@
 
                     if(!cache[id]) {
 
-                        var img = new Image();
+                        var img = new Image(), lowres = false;
 
                         img.onerror = function(){
                             cache[id] = {width:640, height:320};
@@ -360,11 +360,22 @@
                         };
 
                         img.onload = function(){
-                            cache[id] = {width:img.width, height:img.height};
-                            resolve(id, img.width, img.height);
+                            //youtube default 404 thumb, fall back to lowres
+                            if (img.width == 120 && img.height == 90) {
+                                if (!lowres) {
+                                    lowres = true;
+                                    img.src = '//img.youtube.com/vi/' + id + '/0.jpg';
+                                } else {
+                                    cache[id] = {width: 640, height: 320};
+                                    resolve(id, cache[id].width, cache[id].height);
+                                }
+                            } else {
+                                cache[id] = {width: img.width, height: img.height};
+                                resolve(id, img.width, img.height);
+                            }
                         };
 
-                        img.src = '//img.youtube.com/vi/'+id+'/0.jpg';
+                        img.src = '//img.youtube.com/vi/'+id+'/maxresdefault.jpg';
 
                     } else {
                         resolve(id, cache[id].width, cache[id].height);