+++ src/js/components/slider.js
@@ -21,10 +21,13 @@
     UI.component('slider', {
 
         defaults: {
-            center    : false,
-            threshold : 10,
-            infinite  : true,
-            activecls : 'uk-active'
+            center           : false,
+            threshold        : 10,
+            infinite         : true,
+            autoplay         : false,
+            autoplayInterval : 7000,
+            pauseOnHover     : true,
+            activecls        : 'uk-active'
         },
 
         boot:  function() {
@@ -66,65 +69,75 @@
 
                 if ($this.focus == item) return;
 
+                // stop autoplay
+                $this.stop();
+
                 switch(item) {
                     case 'next':
                     case 'previous':
                         $this[item=='next' ? 'next':'previous']();
                         break;
                     default:
-                        $this.updateFocus(parseInt(slide, 10));
+                        $this.updateFocus(parseInt(item, 10));
                 }
             });
 
-            this.container.on('touchstart mousedown', function(evt) {
+            this.container.on({
+                'touchstart mousedown': function(evt) {
 
-                if (evt.originalEvent && evt.originalEvent.touches) {
-                    evt = evt.originalEvent.touches[0];
-                }
-
-                // ignore right click button
-                if (evt.button && evt.button==2 || !$this.active) {
-                    return;
-                }
-
-                anchor  = UI.$(evt.target).is('a') ? UI.$(evt.target) : UI.$(evt.target).parents('a:first');
-                dragged = false;
-
-                if (anchor.length) {
-
-                    anchor.one('click', function(e){
-                        if (dragged) e.preventDefault();
-                    });
-                }
-
-                delayIdle = function(e) {
-
-                    dragged  = true;
-                    dragging = $this;
-                    store    = {
-                        touchx : parseInt(e.pageX, 10),
-                        dir    : 1,
-                        focus  : $this.focus,
-                        base   : $this.options.center ? 'center':'area'
-                    };
-
-                    if (e.originalEvent && e.originalEvent.touches) {
-                        e = e.originalEvent.touches[0];
+                    if (evt.originalEvent && evt.originalEvent.touches) {
+                        evt = evt.originalEvent.touches[0];
                     }
 
-                    dragging.element.data({
-                        'pointer-start': {x: parseInt(e.pageX, 10), y: parseInt(e.pageY, 10)},
-                        'pointer-pos-start': $this.pos
-                    });
+                    // ignore right click button
+                    if (evt.button && evt.button==2 || !$this.active) {
+                        return;
+                    }
 
-                    $this.container.addClass('uk-drag');
+                    // stop autoplay
+                    $this.stop();
 
-                    delayIdle = false;
-                };
+                    anchor  = UI.$(evt.target).is('a') ? UI.$(evt.target) : UI.$(evt.target).parents('a:first');
+                    dragged = false;
 
-                delayIdle.x         = parseInt(evt.pageX, 10);
-                delayIdle.threshold = $this.options.threshold;
+                    if (anchor.length) {
 
+                        anchor.one('click', function(e){
+                            if (dragged) e.preventDefault();
+                        });
+                    }
+
+                    delayIdle = function(e) {
+
+                        dragged  = true;
+                        dragging = $this;
+                        store    = {
+                            touchx : parseInt(e.pageX, 10),
+                            dir    : 1,
+                            focus  : $this.focus,
+                            base   : $this.options.center ? 'center':'area'
+                        };
+
+                        if (e.originalEvent && e.originalEvent.touches) {
+                            e = e.originalEvent.touches[0];
+                        }
+
+                        dragging.element.data({
+                            'pointer-start': {x: parseInt(e.pageX, 10), y: parseInt(e.pageY, 10)},
+                            'pointer-pos-start': $this.pos
+                        });
+
+                        $this.container.addClass('uk-drag');
+
+                        delayIdle = false;
+                    };
+
+                    delayIdle.x         = parseInt(evt.pageX, 10);
+                    delayIdle.threshold = $this.options.threshold;
+
+                },
+                mouseenter: function() { if ($this.options.pauseOnHover) $this.hovering = true;  },
+                mouseleave: function() { $this.hovering = false; }
             });
 
             this.resize(true);
@@ -137,6 +150,12 @@
 
             // prevent dragging links + images
             this.element.find('a,img').attr('draggable', 'false');
+
+            // Set autoplay
+            if (this.options.autoplay) {
+                this.start();
+            }
+
         },
 
         resize: function(focus) {
@@ -187,7 +206,7 @@
                 'transform': ''
             });
 
-            this.updateFocus(0);
+            if (focus) this.updateFocus(this.focus);
         },
 
         updatePos: function(pos) {
@@ -207,7 +226,7 @@
 
             dir = dir || (idx > this.focus ? 1:-1);
 
-            var $this = this, item = this.items.eq(idx), area, i;
+            var item = this.items.eq(idx), area, i;
 
             if (this.options.infinite) {
                 this.infinite(idx, dir);
@@ -280,9 +299,25 @@
             this.updateFocus(focus, -1);
         },
 
+        start: function() {
+
+            this.stop();
+
+            var $this = this;
+
+            this.interval = setInterval(function() {
+                if (!$this.hovering) $this.next();
+            }, this.options.autoplayInterval);
+
+        },
+
+        stop: function() {
+            if (this.interval) clearInterval(this.interval);
+        },
+
         infinite: function(baseidx, direction) {
 
-            var $this = this, item = this.items.eq(baseidx), i, z = baseidx, move = [], lastvisible, area = 0;
+            var $this = this, item = this.items.eq(baseidx), i, z = baseidx, move = [], area = 0;
 
             if (direction == 1) {
 
@@ -439,7 +474,10 @@
 
             dragging.container.removeClass('uk-drag');
 
-            var item  = dragging.items.eq(store.focus), itm, focus = false, i, z;
+            // TODO is this needed?
+            dragging.items.eq(store.focus);
+
+            var itm, focus = false, i, z;
 
             if (store.dir == 1) {