+++ src/js/components/slideshow.js
@@ -39,7 +39,7 @@
                 'uk-animation-top-center',
                 '', // middle-center
                 'uk-animation-bottom-right'
-            ],
+            ]
         },
 
         current  : false,
@@ -56,7 +56,7 @@
                     var slideshow = UI.$(this);
 
                     if (!slideshow.data("slideshow")) {
-                        var obj = UI.slideshow(slideshow, UI.Utils.options(slideshow.attr("data-uk-slideshow")));
+                        UI.slideshow(slideshow, UI.Utils.options(slideshow.attr("data-uk-slideshow")));
                     }
                 });
             });
@@ -256,7 +256,7 @@
 
             if (this.container.hasClass('uk-slideshow-fullscreen')) return;
 
-            var $this = this, height = this.options.height;
+            var height = this.options.height;
 
             if (this.options.height === 'auto') {