+++ CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+### 2.22.0 (August 24, 2015)
+
+  - Added dropdown overlay mode
+  - Added autoplay to slider component
+  - Added animation duration variables to sticky
+  - Added showup + clsinactive option to sticky component
+  - Added swiping option to switcher component
+  - Updated icon component to FontAwesome to 4.4.0
+  - Updated placeholder default variables for all themes
+  - Updated position utility classes
+  - Updated sortable and nestable empty selectors
+  - Updated panel teaser selector
+  - Updated top parameter to support vh values + elements in sticky component
+  - Updated adjust document scrolling while dragging in sortable component
+  - Updated Youtube lighbox plugin to load maxres image
+  - Moved block padding adjustment into themes
+  - Unified sortable and nestable change event parameters
+  - Set Http-Accept-Header if JSON is requested (upload)
+  - Fixed element swapping between grouped lists
+  - Fixed grid match items to wrap onto multiple lines
+  - Fixed dropdown hide behavior
+  - Fixed dropdown prevent show() on active element
+  - Fixed stack margin with dynamic added children
+  - Fixed form icon margin
+  - Fixed timpicker flickering on blur after time picking
+  - Fixed pagination "currentPage" option
+  - Fixed initial focus on confirm dialog
+  - Fixed missing debounced redraw in htmleditor
+  - Fixed body width calculation on active offcanvas
+
+  **Docs and Repo**
+
+  - Added new video tutorial section
+  - Updated some test and docs
+
 ### 2.21.0 (June 01, 2015)
 
   - IMPORTANT: Reworked nestable component